requirements.txt
python-dotenv==1.0.0
streamlit==1.22.0
unstructured==0.9.2
tiktoken==0.4.0
faiss-cpu==1.7.4
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14
langchain_community
tiktoken
langchain-groq
langchainhub
langchain
langgraph
langchain_huggingface
wikipedia
crawl4ai
groq
sentence-transformers
arxiv
rank_bm25
pdfplumber
gradio
crawl4ai
nest_asyncio